refactor some directions
